Revert "OAth token is saved into sessions and should save time on petfinder API"
