+++ controllers/petAPIController.js
@@ -8,7 +8,7 @@ const shelterAPI = require("../utils/shelterAPI/API.js");
 
 //route to return a single pet by id provided by petfinder
 router.get("/pets/:petId", ({ params: { petId } }, res) => {
-    petAPIbyId(petId, { session: { user: { token } } })
+    petAPIbyId(petId)
         .then(petResults => {
             res.json(petResults)
             console.log("pet by id", petResults);
@@ -21,8 +21,9 @@ router.get("/pets/:petId", ({ params: { petId } }, res) => {
 
 //route to get array of animals by user preferences - should hopefully work once we have a sessions object
 
+// UNCOMMENT FOR LIVE DATA
 router.get("/pets/", ({ session: { user: { postcode, hasCats, hasDogs, hasKids, whichSpecies } } }, res) => {
-    console.log(postcode, hasDogs, hasKids, hasCats, whichSpecies, token)
+    console.log(postcode, hasDogs, hasKids, hasCats, whichSpecies)
     petAPI(postcode, hasDogs, hasKids, hasCats, whichSpecies)
         .then(petResults => {
             res.json(petResults)
@@ -31,5 +32,16 @@ router.get("/pets/", ({ session: { user: { postcode, hasCats, hasDogs, hasKids, 
             res.status(500).end()
         })
 })
+
+// ROUTE FOR BASIC TESTING
+// router.get("/pets/", (req, res) => {
+//     petAPI()
+//         .then(petResults => {
+//             res.json(petResults)
+//         }).catch(err => {
+//             console.log(err);
+//             res.status(500).end()
+//         })
+// })
 
 module.exports = router;
