+++ controllers/usersController.js
@@ -1,7 +1,6 @@
 const router = require("express").Router();
 const db = require("../models");
-const bcrypt = require('bcrypt');
-const getToken = require("../utils/petAPI/getToken.js");
+const bcrypt = require('bcrypt')
 
 //BASE URL FOR ALL ROUTES ON THIS PAGE: /api/users
 
@@ -10,13 +9,13 @@ const getToken = require("../utils/petAPI/getToken.js");
 //TODO: Remove or comment out on official deployment for security
 router.get("/userlist/", (req, res) => {
     db.User.findAll({})
-        .then(userList => {
-            res.json(userList);
-        })
-        .catch(err => {
-            console.log(err);
-            res.status(500).end()
-        })
+    .then(userList => {
+        res.json(userList);
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).end()
+    })
 })
 
 // CHANGED ROUTE SO THAT THE OTHER ROUTES WOULD NOT HIT THIS ROUTE BY ACCIDENT.
@@ -94,14 +93,14 @@ router.post('/login', (req, res) => {
         where: {
             email: req.body.email
         }
-    }).then(async user => {
+    }).then(user => {
         if (!user) {
             res.status(404).send("No such user exists");
         } else {
             if (
+                // UNCOMMENT WHEN YOU WANT TO AUTHENTICATE.
                 bcrypt.compareSync
                     (req.body.password, user.password)) {
-                const { data: { access_token } } = await getToken()
                 req.session.user = {
                     firstName: user.firstName,
                     lastName: user.lastName,
@@ -111,8 +110,7 @@ router.post('/login', (req, res) => {
                     hasKids: user.hasKids,
                     hasCats: user.hasCats,
                     hasDogs: user.hasDogs,
-                    whichSpecies: user.whichSpecies,
-                    token: access_token
+                    whichSpecies: user.whichSpecies
                 }
                 res.json(req.session);
             } else {
